# specs/006-phase2-fullstack-webapp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API - Phase II
  description: |
    RESTful API for the Evolution of Todo Full-Stack Web Application.
    All task endpoints require JWT authentication via Better Auth.
  version: 1.0.0
  contact:
    name: Hackathon II Team

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://your-backend.railway.app
    description: Production (Railway)

tags:
  - name: Tasks
    description: Task CRUD operations
  - name: Health
    description: Health check endpoints

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/{user_id}/tasks:
    parameters:
      - $ref: '#/components/parameters/UserId'

    get:
      tags: [Tasks]
      summary: List all tasks for user
      operationId: listTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Tasks]
      summary: Create a new task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/TaskId'

    get:
      tags: [Tasks]
      summary: Get a specific task
      operationId: getTask
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update a task
      operationId: updateTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Task deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/TaskId'

    patch:
      tags: [Tasks]
      summary: Toggle task completion status
      operationId: toggleTaskComplete
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: User ID (must match authenticated user)
      schema:
        type: string
        example: usr_abc123

    TaskId:
      name: task_id
      in: path
      required: true
      description: Task ID
      schema:
        type: integer
        example: 1

  schemas:
    Task:
      type: object
      required: [id, user_id, title, completed, created_at, updated_at]
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        user_id:
          type: string
          description: Owner user ID
          example: usr_abc123
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional task description
          example: Milk, eggs, bread
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: '2025-12-29T10:00:00Z'
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: '2025-12-29T10:00:00Z'

    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          description: Optional task description
          example: Milk, eggs, bread

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: New task title
          example: Buy groceries and fruits
        description:
          type: string
          maxLength: 1000
          nullable: true
          description: New task description
          example: Milk, eggs, bread, apples

    TaskList:
      type: object
      required: [tasks, total]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        total:
          type: integer
          description: Total number of tasks
          example: 5

    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
          example: Task not found

  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Title is required

    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Invalid token

    Forbidden:
      description: User does not own this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Access denied

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Task not found